requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.2.1
Django==4.1.7
django-dynamic-fixture==3.1.2
django-extensions==3.2.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==6.0.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.0
phonenumbers==8.13.6
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-flake8==1.1.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
